Struct vkgen::Array

pub struct Array {
    pub ty: ArrayType,
    pub size: Option<ArraySize>,
    pub null_terminate: Option<bool>,
    pub c_size: Option<CExpression>,
}

C array (dynamic or static)

Static arrays have ty set to ArrayType::Static and the sizing attribute.

Dynamic arrays have ty set to ArrayType::Dynamic, and may have the null-terminate attribute, size attribute, or both. If it has neither, then the size is not determinable. If it has both, then the null-terminate size refers to a char* of some form, with the size referring to the outer array.

Fields
ty: ArrayType size: Option<ArraySize>

The size of the array. The value of size is represented differently depending on the array type.

If the array is "static", size must be present. Static sizes can be integer constants or enumerator references.

If the array is "dynamic", size may be present. If not present, then the dynamic array is either null-terminated or has no specific size. If the construct is a pointer-pointer, then this only specifies the size of the outer array.

null_terminate: Option<bool>

If true, then a string variable (some form of pointer with the type char) is null-terminated. If the type is a pointer-pointer or a pointer-const-pointer, then the null-terminator refers only to the innermost string pointers.

c_size: Option<CExpression>

A C expression that specifies the number of elements in a static or dynamic array.