Struct vkgen::Command

pub struct Command {
    pub name: Identifier,
    pub notation: Option<String>,
    pub return_type: ReturnType,
    pub param: Vec<Field>,
    pub external_sync: Option<ExternalSync>,
    pub renderpass: Option<Renderpass>,
    pub cmdbufferlevel: Option<CommaSeparatedIdentifiers>,
    pub pipeline: Option<Pipeline>,
    pub queues: Option<CommaSeparatedIdentifiers>,
}

Defines a single Vulkan entrypoint.

Fields
name: Identifier notation: Option<String> return_type: ReturnType param: Vec<Field> external_sync: Option<ExternalSync> renderpass: Option<Renderpass>

Specifies how the command may be called, relative to render pass instance scopes. When not specified, the command does not care about render pass scoping, or the command isn't a vkCmd command.

cmdbufferlevel: Option<CommaSeparatedIdentifiers>

Specifies the level of command buffer that the command can be used with. When not specified, the command does not go into a command buffer. { "primary" | "secondary" | "primary,secondary" }

pipeline: Option<Pipeline>

Specifies the kind of pipeline that a command can be used with.

queues: Option<CommaSeparatedIdentifiers>

Comma-separated list of queues that support this operation. When not specified, the function cannot be used in a queue.