Struct vkgen::ExtensionSpecification

pub struct ExtensionSpecification {
    pub profile: Option<String>,
    pub notation: Option<String>,
    pub extension: Option<Identifier>,
    pub api: Option<String>,
    pub elements: Vec<ExtensionSpecificationElement>,
}

Fields
profile: Option<String>

The name of the API profile which this specialization element is associated. All specified requirements/exclusions apply only to that profile.

notation: Option<String> extension: Option<Identifier>

Specifies the name of an extension which this require statement needs for its inclusions to be imported.

api: Option<String>

Specifies that the extensions or removals only apply to that specific API being generated. If not present, it applies to all APIs.

elements: Vec<ExtensionSpecificationElement>